Доработка issues
