+++ libs/api/shared/prisma/seed.ts
@@ -2,6 +2,7 @@ import 'dotenv/config';
 import * as crypto from 'node:crypto';
 import { PrismaPg } from '@prisma/adapter-pg';
 import * as bcrypt from 'bcryptjs';
+import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import {
   AssessmentStatus,
   DocumentType,
@@ -27,6 +28,107 @@ if (!connectionString) {
 const prisma = new PrismaClient({
   adapter: new PrismaPg({ connectionString }),
 });
+
+const s3 = new S3Client({
+  region: process.env['S3_REGION'] ?? 'us-east-1',
+  endpoint: process.env['S3_ENDPOINT'],
+  credentials: {
+    accessKeyId: process.env['S3_ACCESS_KEY'] ?? 'minioadmin',
+    secretAccessKey: process.env['S3_SECRET_KEY'] ?? 'minioadmin',
+  },
+  forcePathStyle: process.env['S3_FORCE_PATH_STYLE'] !== 'false',
+});
+const S3_BUCKET = process.env['S3_BUCKET_PAYMENT_DOCUMENTS'] ?? 'payment-documents';
+
+function makeMinimalPdf(): Buffer {
+  return Buffer.from(
+    '%PDF-1.4\n1 0 obj<</Type/Catalog/Pages 2 0 R>>endobj 2 0 obj<</Type/Pages/Kids[3 0 R]/Count 1>>endobj 3 0 obj<</Type/Page/MediaBox[0 0 200 200]/Parent 2 0 R>>endobj\nxref\n0 4\n0000000000 65535 f \n0000000009 00000 n \n0000000058 00000 n \n0000000115 00000 n \ntrailer<</Size 4/Root 1 0 R>>\nstartxref\n190\n%%EOF',
+  );
+}
+
+function escapeHtml(value: string): string {
+  return value
+    .replaceAll('&', '&amp;')
+    .replaceAll('<', '&lt;')
+    .replaceAll('>', '&gt;')
+    .replaceAll('"', '&quot;')
+    .replaceAll("'", '&#39;');
+}
+
+function formatMoney(value: string): string {
+  const amount = Number(value);
+  if (Number.isNaN(amount)) return value;
+  return new Intl.NumberFormat('ru-RU', {
+    style: 'currency',
+    currency: 'RUB',
+    minimumFractionDigits: 2,
+    maximumFractionDigits: 2,
+  }).format(amount);
+}
+
+function makeReceiptHtml(params: {
+  paymentId: string;
+  amount: string;
+  paymentDate: string | Date;
+  paymentMethod: string;
+  transactionId: string;
+  type: string;
+}): Buffer {
+  const dateStr =
+    params.paymentDate instanceof Date
+      ? new Intl.DateTimeFormat('ru-RU', { dateStyle: 'medium', timeStyle: 'short' }).format(
+          params.paymentDate,
+        )
+      : params.paymentDate;
+
+  const typeLabel =
+    params.type === 'Subscription'
+      ? 'Оплата подписки'
+      : params.type === 'DocumentCopy'
+        ? 'Оплата копии документа'
+        : 'Оценка недвижимости';
+
+  const html = `<!doctype html>
+<html lang="ru">
+<head><meta charset="utf-8"/><title>Чек оплаты</title>
+<style>
+  :root{font-family:-apple-system,BlinkMacSystemFont,'Segoe UI',sans-serif;line-height:1.5}
+  body{margin:0;padding:32px;background:#f3f5f7;color:#102030}
+  .receipt{max-width:760px;margin:0 auto;background:#fff;border-radius:20px;padding:32px;box-shadow:0 16px 48px rgba(16,32,48,.12)}
+  .title{margin:0 0 8px;font-size:28px;line-height:1.1}
+  .subtitle{margin:0;color:#556677}
+  .grid{display:grid;grid-template-columns:repeat(2,1fr);gap:12px;margin:24px 0}
+  .card{padding:16px;border-radius:16px;background:#f8fafc;border:1px solid #e4ebf2}
+  .label{display:block;margin-bottom:6px;color:#64748b;font-size:12px;text-transform:uppercase;letter-spacing:.04em}
+  .value{font-size:16px;font-weight:600;word-break:break-word}
+  table{width:100%;border-collapse:collapse;border-radius:16px;border:1px solid #e4ebf2}
+  th,td{padding:14px 16px;text-align:left;border-bottom:1px solid #e4ebf2}
+  th{background:#f8fafc;color:#64748b;font-size:12px;text-transform:uppercase}
+  tr:last-child td{border-bottom:none}
+  .summary{margin-top:20px;display:flex;justify-content:flex-end}
+  .summary-card{min-width:240px;padding:16px 20px;border-radius:16px;background:#102030;color:#fff}
+  .summary-card strong{display:block;margin-top:6px;font-size:24px}
+  .footer{margin-top:24px;padding-top:16px;border-top:1px solid #e4ebf2}
+  .note{margin:0;color:#556677}
+  @media print{body{padding:0;background:#fff}.receipt{max-width:none;border-radius:0;box-shadow:none}}
+</style></head>
+<body><main class="receipt">
+  <h1 class="title">Чек оплаты</h1>
+  <p class="subtitle">Документ по подтвержденной оплате в системе.</p>
+  <section class="grid">
+    <article class="card"><span class="label">ID платежа</span><span class="value">${escapeHtml(params.paymentId)}</span></article>
+    <article class="card"><span class="label">ID транзакции</span><span class="value">${escapeHtml(params.transactionId)}</span></article>
+    <article class="card"><span class="label">Дата оплаты</span><span class="value">${escapeHtml(dateStr)}</span></article>
+    <article class="card"><span class="label">Способ оплаты</span><span class="value">${escapeHtml(params.paymentMethod)}</span></article>
+  </section>
+  <table><thead><tr><th>Описание</th><th>Кол-во</th><th>Сумма</th></tr></thead>
+  <tbody><tr><td>${escapeHtml(typeLabel)}</td><td>1</td><td>${escapeHtml(formatMoney(params.amount))}</td></tr></tbody></table>
+  <section class="summary"><div class="summary-card">Итого<strong>${escapeHtml(formatMoney(params.amount))}</strong></div></section>
+  <footer class="footer"><p class="note">Документ сформирован автоматически. Seed-данные для тестирования.</p></footer>
+</main></body></html>`;
+
+  return Buffer.from(html, 'utf8');
+}
 
 const SEED_COUNT = Number(process.env['SEED_SIZE'] ?? 100);
 const SEED_USERS_PER_ROLE = 10;
@@ -513,6 +615,43 @@ async function upsertPayments(
       status = statuses[idx % 4];
     }
 
+    const receiptObjectKey = `payment-documents/${id}/${crypto.randomUUID()}.html`;
+    const receiptFileName = `receipt-${i}.html`;
+    const receiptStatus =
+      status === PaymentStatus.Completed
+        ? PaymentReceiptStatus.Available
+        : status === PaymentStatus.Failed
+          ? PaymentReceiptStatus.Failed
+          : PaymentReceiptStatus.Pending;
+
+    // Upload a receipt HTML with real payment data to MinIO for completed payments
+    if (receiptStatus === PaymentReceiptStatus.Available) {
+      try {
+        await s3.send(
+          new PutObjectCommand({
+            Bucket: S3_BUCKET,
+            Key: receiptObjectKey,
+            Body: makeReceiptHtml({
+              paymentId: id,
+              amount,
+              paymentDate,
+              paymentMethod: methods[i % 3],
+              transactionId: `TXN-SEED-${pad(i, 5)}`,
+              type:
+                type === PaymentType.Subscription
+                  ? 'Subscription'
+                  : type === PaymentType.DocumentCopy
+                    ? 'DocumentCopy'
+                    : 'Assessment',
+            }),
+            ContentType: 'text/html; charset=utf-8',
+          }),
+        );
+      } catch (err) {
+        console.warn(`Seed: failed to upload receipt for payment ${id} to S3`, err);
+      }
+    }
+
     await prisma.payment.upsert({
       where: { id },
       update: {
@@ -527,9 +666,11 @@ async function upsertPayments(
         status,
         paymentMethod: methods[i % 3],
         transactionId: `TXN-SEED-${pad(i, 5)}`,
-        attachmentFileName: `receipt-${i}.pdf`,
-        attachmentFileUrl: `https://example.local/files/receipt-${i}.pdf`,
-        receiptStatus: PaymentReceiptStatus.Available,
+        attachmentFileName:
+          receiptStatus === PaymentReceiptStatus.Available ? receiptFileName : null,
+        attachmentFileUrl:
+          receiptStatus === PaymentReceiptStatus.Available ? receiptObjectKey : null,
+        receiptStatus,
       },
       create: {
         id,
@@ -544,9 +685,11 @@ async function upsertPayments(
         status,
         paymentMethod: methods[i % 3],
         transactionId: `TXN-SEED-${pad(i, 5)}`,
-        attachmentFileName: `receipt-${i}.pdf`,
-        attachmentFileUrl: `https://example.local/files/receipt-${i}.pdf`,
-        receiptStatus: PaymentReceiptStatus.Available,
+        attachmentFileName:
+          receiptStatus === PaymentReceiptStatus.Available ? receiptFileName : null,
+        attachmentFileUrl:
+          receiptStatus === PaymentReceiptStatus.Available ? receiptObjectKey : null,
+        receiptStatus,
       },
     });
   }
