+++ libs/web/admin/src/lib/admin.routes.ts
@@ -2,7 +2,6 @@ import { Route } from '@angular/router';
 import { Admin } from './admin/admin';
 import { PlaceholderPageRoute } from '@notary-portal/ui';
 import { Payments } from './features/payments/payments';
-import { Applications } from './features/applications/applications';
 import { PaymentFormComponent } from './features/payments/payment-form.component';
 
 const placeholder = (title: string, features: string[]): Partial<Route> => ({
@@ -17,36 +16,10 @@ export const adminRoutes: Route[] = [
     children: [
       { path: '', ...placeholder('Главное меню', ['Обзор панели администратора']) } as Route,
       {
-        path: 'users',
+        path: 'applications',
         loadComponent: () =>
           import('./features/RequestAssessment/RequestAssessment').then((m) => m.RequestAssessment),
-      } as Route,
-      {
-        path: 'orders',
-        loadComponent: () =>
-          import('./features/RequestAssessment/requests/requests').then((m) => m.RequestsComponent),
-      } as Route,
-      {
-        path: 'orders/statuses',
-        ...placeholder('Управление статусами', [
-          'Изменение статусов заказов',
-          'Отслеживание переходов между этапами',
-        ]),
-      } as Route,
-      {
-        path: 'orders/queue',
-        ...placeholder('Очередь оценок', [
-          'Список заявок, ожидающих оценки',
-          'Распределение по нотариусам',
-        ]),
-      } as Route,
-      {
-        path: 'orders/moderation',
-        ...placeholder('Ручная модерация', [
-          'Проверка и модерация спорных заявок',
-          'Ручные корректировки',
-        ]),
-      } as Route,
+      },
       {
         path: 'payments/new',
         component: PaymentFormComponent,
@@ -60,10 +33,6 @@ export const adminRoutes: Route[] = [
         component: Payments,
       },
 
-      {
-        path: 'applications',
-        component: Applications,
-      },
       {
         path: 'subscriptions',
         ...placeholder('Подписки', ['Просмотр списка подписок']),
