+++ libs/web/admin/src/lib/admin/admin.ts
@@ -2,12 +2,12 @@ import { Component, ViewEncapsulation, inject } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { DashboardLayout } from '@notary-portal/ui';
 import { AdminPaymentsApiService } from '../features/payments/payments-api.service';
-import { AdminApplicationsApiService } from '../features/applications/applications-api.service';
+import { AdminApplicationsApiService } from '../features/RequestAssessment/applications-api.service';
 
 const ADMIN_MENU = [
   { label: 'Главное меню', route: '.', icon: '☰', exact: true },
   { label: 'Пользователи', route: 'users', icon: '👥' },
-  { label: 'Управление заказами', route: 'orders', icon: '📄', exact: true },
+  { label: 'Управление заказами', route: 'applications', icon: '📄' },
   { label: 'Управление статусами', route: 'orders/statuses', icon: '🔄' },
   { label: 'Очередь оценок', route: 'orders/queue', icon: '📝' },
   { label: 'Ручная модерация', route: 'orders/moderation', icon: '✅' },
