+++ libs/web/admin/src/lib/features/RequestAssessment/RequestAssessment.spec.ts
@@ -1,5 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { provideRouter } from '@angular/router';
 import { RequestAssessment } from './RequestAssessment';
+import { AdminApplicationsApiService } from './applications-api.service';
 
 describe('RequestAssessment', () => {
   let component: RequestAssessment;
@@ -8,6 +10,15 @@ describe('RequestAssessment', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [RequestAssessment],
+      providers: [
+        provideRouter([]),
+        {
+          provide: AdminApplicationsApiService,
+          useValue: {
+            getAllApplications: async () => [],
+          },
+        },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(RequestAssessment);
