+++ libs/web/admin/src/lib/features/RequestAssessment/RequestAssessment.ts
@@ -1,297 +1,409 @@
-import {
-  ChangeDetectorRef,
-  Component,
-  HostListener,
-  inject,
-  OnDestroy,
-  OnInit,
-} from '@angular/core';
+import { Component, HostListener, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { Subject, Subscription } from 'rxjs';
-import { debounceTime } from 'rxjs/operators';
+import { Router, ActivatedRoute } from '@angular/router';
+import { inject } from '@angular/core';
+import { buildRpcBaseUrl, TokenStore } from '@notary-portal/ui';
+import { Subscription } from 'rxjs';
+import {
+  PAYMENT_METHOD_LABELS,
+  PAYMENT_STATUS_LABELS,
+  PAYMENT_TYPE_LABELS,
+  PaymentMethod,
+  PaymentStatus,
+  PaymentType,
+} from '../payments/payments.shared';
+import { AdminApplicationsApiService } from './applications-api.service';
 
-export interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  middleName: string;
-  email: string;
-  phoneNumber: string;
-  role: 'Applicant' | 'Notary' | 'Admin';
-  subscriptionPlan: 'Basic' | 'Premium' | 'Enterprise';
-  isActive: boolean;
-  createdAt: string;
-  updatedAt: string;
+export interface Application {
+  id: string | number;
+  date: string;
+  sender: string;
+  recipient: string;
+  type: PaymentType;
+  amount: number;
+  fee: number;
+  paymentMethod?: PaymentMethod;
+  transactionId?: string;
+  receiptFileName?: string;
+  receiptFileUrl?: string;
+  statementId?: string;
+  status: PaymentStatus;
+  statusText: string;
 }
 
-type UserFilterColumn =
-  | 'fullName'
-  | 'email'
-  | 'phoneNumber'
-  | 'role'
-  | 'isActive'
-  | 'createdAt'
-  | 'actions';
+type PaginationItem = number | 'ellipsis';
 
 @Component({
   selector: 'lib-request-assessment',
+  standalone: true,
   imports: [CommonModule, FormsModule],
   templateUrl: './RequestAssessment.html',
   styleUrl: './RequestAssessment.scss',
 })
 export class RequestAssessment implements OnInit, OnDestroy {
-  private cdr = inject(ChangeDetectorRef);
-
-  currentView: 'list' | 'detail' = 'list';
-
-  users: User[] = [];
-  filteredUsers: User[] = [];
-  paginatedUsers: User[] = [];
+  applications: Application[] = [];
+  loading = true;
+  loadError: string | null = null;
 
   searchTerm = '';
-  private searchSubject$ = new Subject<string>();
-  private searchSubscription?: Subscription;
-
-  readonly headerColumns: { key: UserFilterColumn; label: string }[] = [
-    { key: 'fullName', label: 'ФИО' },
-    { key: 'email', label: 'Email' },
-    { key: 'phoneNumber', label: 'Телефон' },
-    { key: 'role', label: 'Роль' },
-    { key: 'isActive', label: 'Статус' },
-    { key: 'createdAt', label: 'Дата регистрации' },
+  statusFilter: '' | PaymentStatus = '';
+  readonly headerColumns: { key: FilterColumn; label: string }[] = [
+    { key: 'id', label: 'ID' },
+    { key: 'date', label: 'Дата' },
+    { key: 'sender', label: 'Отправитель' },
+    { key: 'recipient', label: 'Получатель' },
+    { key: 'type', label: 'Тип' },
+    { key: 'amount', label: 'Сумма' },
+    { key: 'fee', label: 'Комиссия' },
+    { key: 'paymentMethod', label: 'Метод оплаты' },
+    { key: 'transactionId', label: 'ID транзакции' },
+    { key: 'receipt', label: 'Чек' },
+    { key: 'statement', label: 'Заявление' },
+    { key: 'status', label: 'Статус' },
     { key: 'actions', label: 'Действия' },
   ];
 
-  activeFilterColumn: UserFilterColumn | null = null;
-  columnSelectedValues: Record<UserFilterColumn, string[]> = {
-    fullName: [],
-    email: [],
-    phoneNumber: [],
-    role: [],
-    isActive: [],
-    createdAt: [],
+  activeFilterColumn: FilterColumn | null = null;
+  columnSelectedValues: Record<FilterColumn, string[]> = {
+    id: [],
+    date: [],
+    sender: [],
+    recipient: [],
+    type: [],
+    amount: [],
+    fee: [],
+    paymentMethod: [],
+    transactionId: [],
+    receipt: [],
+    statement: [],
+    status: [],
     actions: [],
   };
-  currentSortColumn: UserFilterColumn | null = null;
+
+  currentSortColumn: FilterColumn | null = null;
   currentSortDirection: '' | 'asc' | 'desc' = '';
   filterSearch = '';
   filterSortDraft: '' | 'asc' | 'desc' = '';
   filterSelectedDraft = new Set<string>();
 
+  filterMenuStyle: Record<string, string> = {};
+
+  readonly today: string = new Date().toISOString().split('T')[0];
+
+  pageSize = 7;
   currentPage = 1;
-  usersPerPage = 10;
-  totalPages = 0;
-  pages: number[] = [];
+  readonly skeletonRows = Array.from({ length: 6 }, (_, index) => index);
 
-  selectedUser: User | null = null;
+  isCreateEditModalOpen = false;
+  isViewModalOpen = false;
+  isDeleteModalOpen = false;
+  isEditMode = false;
 
-  showUserModal = false;
-  isEditing = false;
-  formData: Partial<User> = {};
-  formErrors: Record<string, string> = {};
+  currentApplication: Application = this.resetApplication();
 
-  showDeleteModal = false;
-  userToDelete: User | null = null;
+  private router = inject(Router);
+  private route = inject(ActivatedRoute);
+  private readonly api = inject(AdminApplicationsApiService);
+  private readonly tokenStore = inject(TokenStore);
+  private dataSub?: Subscription;
 
-  readonly roleLabels: Record<string, string> = {
-    Applicant: 'Заявитель',
-    Notary: 'Нотариус',
-    Admin: 'Администратор',
-  };
+  async openReceipt(applicationId: string | number): Promise<void> {
+    const token = this.tokenStore.getAccessToken();
+    if (!token) return;
 
-  readonly roleBadgeClasses: Record<string, string> = {
-    Applicant: 'badge-primary',
-    Notary: 'badge-warning',
-    Admin: 'badge-info',
-  };
+    try {
+      const response = await fetch(`${buildRpcBaseUrl()}/api/payments/${applicationId}/receipt`, {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+      if (!response.ok) return;
+
+      const blob = await response.blob();
+      const objectUrl = URL.createObjectURL(blob);
+      window.open(objectUrl, '_blank', 'noopener,noreferrer');
+      setTimeout(() => URL.revokeObjectURL(objectUrl), 60_000);
+    } catch {
+      // ignore
+    }
+  }
+
+  async downloadReceipt(applicationId: string | number, fileName?: string): Promise<void> {
+    const token = this.tokenStore.getAccessToken();
+    if (!token) return;
+
+    try {
+      const response = await fetch(`${buildRpcBaseUrl()}/api/payments/${applicationId}/receipt`, {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+      if (!response.ok) return;
+
+      const blob = await response.blob();
+      const objectUrl = URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.href = objectUrl;
+      a.download = fileName || `receipt-${applicationId}.pdf`;
+      document.body.appendChild(a);
+      a.click();
+      document.body.removeChild(a);
+      setTimeout(() => URL.revokeObjectURL(objectUrl), 60_000);
+    } catch {
+      // ignore
+    }
+  }
+
+  constructor() {
+    this.route.queryParams.subscribe((params) => {
+      const id = params['assessmentId'];
+      if (id) this.searchTerm = String(id);
+    });
+  }
 
   ngOnInit(): void {
-    this.searchSubscription = this.searchSubject$.pipe(debounceTime(300)).subscribe(() => {
-      this.applyFilters();
-      this.cdr.detectChanges();
+    this.loading = true;
+    this.loadError = null;
+    this.api.preload();
+    this.dataSub = this.api.applications$.subscribe({
+      next: (data) => {
+        if (data !== null) {
+          this.applications = data;
+          this.loading = false;
+        }
+      },
+      error: (err) => {
+        console.error('Failed to load admin applications', err);
+        this.applications = [];
+        this.loadError = 'Не удалось загрузить данные заявок с сервера';
+        this.loading = false;
+      },
     });
-    this.initializeData();
-    this.applyFilters();
   }
 
   ngOnDestroy(): void {
-    this.searchSubscription?.unsubscribe();
+    this.dataSub?.unsubscribe();
   }
 
-  onSearchChange(value: string): void {
-    this.searchTerm = value;
-    if (!value) {
-      this.applyFilters();
-    } else {
-      this.searchSubject$.next(value);
-    }
-  }
+  get filteredApplications(): Application[] {
+    const term = this.searchTerm.trim().toLowerCase();
+    const status = this.statusFilter;
 
-  private generateUUID(): string {
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-      const r = (Math.random() * 16) | 0;
-      const v = c === 'x' ? r : (r & 0x3) | 0x8;
-      return v.toString(16);
+    const filtered = this.applications.filter((a) => {
+      const matchesStatus = !status || a.status === status;
+      const matchesTerm =
+        !term ||
+        a.sender.toLowerCase().includes(term) ||
+        a.recipient.toLowerCase().includes(term) ||
+        a.id.toString().includes(term) ||
+        (a.transactionId?.toLowerCase().includes(term) ?? false) ||
+        (a.statementId?.toLowerCase().includes(term) ?? false);
+
+      const matchesColumnFilters = this.matchesColumnFilters(a);
+
+      return matchesStatus && matchesTerm && matchesColumnFilters;
     });
+
+    return [...filtered].sort((a, b) => this.compareByActiveSort(a, b));
   }
 
-  private initializeData(): void {
-    const existing = localStorage.getItem('users');
-    if (existing) {
-      const parsed = JSON.parse(existing);
-      if (parsed.length > 0 && parsed[0].fullName !== undefined) {
-        localStorage.removeItem('users');
-      }
-    }
-
-    if (!localStorage.getItem('users')) {
-      const testUsers: User[] = [
-        {
-          id: this.generateUUID(),
-          firstName: 'Иван',
-          lastName: 'Иванов',
-          middleName: 'Иванович',
-          email: 'ivanov@example.com',
-          phoneNumber: '+7 (999) 123-45-67',
-          role: 'Applicant',
-          subscriptionPlan: 'Basic',
-          isActive: true,
-          createdAt: '2024-01-15T10:30:00',
-          updatedAt: '2024-01-15T10:30:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Мария',
-          lastName: 'Петрова',
-          middleName: 'Сергеевна',
-          email: 'petrova@example.com',
-          phoneNumber: '+7 (999) 234-56-78',
-          role: 'Notary',
-          subscriptionPlan: 'Premium',
-          isActive: true,
-          createdAt: '2024-01-20T14:20:00',
-          updatedAt: '2024-02-01T09:15:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Алексей',
-          lastName: 'Сидоров',
-          middleName: 'Петрович',
-          email: 'sidorov@example.com',
-          phoneNumber: '+7 (999) 345-67-89',
-          role: 'Admin',
-          subscriptionPlan: 'Enterprise',
-          isActive: true,
-          createdAt: '2024-02-01T11:45:00',
-          updatedAt: '2024-02-01T11:45:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Елена',
-          lastName: 'Козлова',
-          middleName: 'Викторовна',
-          email: 'kozlova@example.com',
-          phoneNumber: '+7 (999) 456-78-90',
-          role: 'Applicant',
-          subscriptionPlan: 'Basic',
-          isActive: false,
-          createdAt: '2024-02-05T16:00:00',
-          updatedAt: '2024-02-10T12:30:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Дмитрий',
-          lastName: 'Новиков',
-          middleName: 'Александрович',
-          email: 'novikov@example.com',
-          phoneNumber: '+7 (999) 567-89-01',
-          role: 'Notary',
-          subscriptionPlan: 'Premium',
-          isActive: true,
-          createdAt: '2024-02-08T09:15:00',
-          updatedAt: '2024-02-08T09:15:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Ольга',
-          lastName: 'Васильева',
-          middleName: 'Михайловна',
-          email: 'vasilieva@example.com',
-          phoneNumber: '+7 (999) 678-90-12',
-          role: 'Applicant',
-          subscriptionPlan: 'Enterprise',
-          isActive: true,
-          createdAt: '2024-02-12T13:45:00',
-          updatedAt: '2024-02-12T13:45:00',
-        },
-        {
-          id: this.generateUUID(),
-          firstName: 'Сергей',
-          lastName: 'Морозов',
-          middleName: 'Владимирович',
-          email: 'morozov@example.com',
-          phoneNumber: '+7 (999) 789-01-23',
-          role: 'Notary',
-          subscriptionPlan: 'Basic',
-          isActive: false,
-          createdAt: '2024-02-15T10:00:00',
-          updatedAt: '2024-02-20T15:20:00',
-        },
-      ];
-      localStorage.setItem('users', JSON.stringify(testUsers));
-    }
-
-    this.users = JSON.parse(localStorage.getItem('users') ?? '[]');
+  get paginatedApplications(): Application[] {
+    const start = (this.currentPage - 1) * this.pageSize;
+    const end = start + this.pageSize;
+    return this.filteredApplications.slice(start, end);
   }
 
-  private saveToStorage(): void {
-    localStorage.setItem('users', JSON.stringify(this.users));
+  get totalPages(): number {
+    return Math.max(1, Math.ceil(this.filteredApplications.length / this.pageSize));
   }
 
-  // ========== ФИЛЬТРАЦИЯ И ПАГИНАЦИЯ ==========
+  get pages(): number[] {
+    return Array.from({ length: this.totalPages }, (_, i) => i + 1);
+  }
 
-  applyFilters(): void {
-    let result = [...this.users];
+  get paginationItems(): PaginationItem[] {
+    const total = this.totalPages;
+    const current = this.currentPage;
 
-    if (this.searchTerm) {
-      const term = this.searchTerm.toLowerCase();
-      result = result.filter(
-        (u) =>
-          this.getFullName(u).toLowerCase().includes(term) || u.email.toLowerCase().includes(term),
-      );
+    if (total <= 7) {
+      return this.pages;
     }
 
-    result = result.filter((u) => this.matchesColumnFilters(u));
-    result.sort((a, b) => this.compareByActiveSort(a, b));
+    const items: PaginationItem[] = [1];
+    const start = Math.max(2, current - 1);
+    const end = Math.min(total - 1, current + 1);
 
-    this.filteredUsers = result;
+    if (start > 2) {
+      items.push('ellipsis');
+    }
+
+    for (let page = start; page <= end; page += 1) {
+      items.push(page);
+    }
+
+    if (end < total - 1) {
+      items.push('ellipsis');
+    }
+
+    items.push(total);
+    return items;
+  }
+
+  isPageItem(page: PaginationItem): page is number {
+    return typeof page === 'number';
+  }
+
+  setPage(page: number): void {
+    if (page < 1 || page > this.totalPages) return;
+    this.currentPage = page;
+  }
+
+  prevPage(): void {
+    this.setPage(this.currentPage - 1);
+  }
+
+  nextPage(): void {
+    this.setPage(this.currentPage + 1);
+  }
+
+  onFiltersChanged(): void {
     this.currentPage = 1;
-    this.updatePagination();
   }
 
-  // ========== COLUMN FILTER ==========
+  private resetApplication(): Application {
+    return {
+      id: 0,
+      date: this.today,
+      sender: '',
+      recipient: '',
+      type: 'Assessment',
+      amount: 0,
+      fee: 0,
+      paymentMethod: 'card',
+      status: 'pending',
+      statusText: PAYMENT_STATUS_LABELS.pending,
+    };
+  }
 
-  toggleColumnFilter(column: UserFilterColumn, event: MouseEvent): void {
+  openCreateModal(): void {
+    this.isEditMode = false;
+    this.currentApplication = this.resetApplication();
+    this.currentApplication.id = this.getNextLocalId();
+    this.isCreateEditModalOpen = true;
+  }
+
+  private getNextLocalId(): number {
+    const numericIds = this.applications
+      .map((application) => Number(application.id))
+      .filter((id) => Number.isFinite(id));
+
+    return Math.max(...numericIds, 0) + 1;
+  }
+
+  openEditModal(app: Application): void {
+    this.isEditMode = true;
+    this.currentApplication = { ...app };
+    this.isCreateEditModalOpen = true;
+  }
+
+  openViewModal(app: Application): void {
+    this.currentApplication = { ...app };
+    this.isViewModalOpen = true;
+  }
+
+  openDeleteModal(app: Application): void {
+    this.currentApplication = { ...app };
+    this.isDeleteModalOpen = true;
+  }
+
+  closeModals(): void {
+    this.isCreateEditModalOpen = false;
+    this.isViewModalOpen = false;
+    this.isDeleteModalOpen = false;
+  }
+
+  onModalBackdropClick(event: MouseEvent): void {
+    if ((event.target as HTMLElement).classList.contains('modal')) {
+      this.closeModals();
+    }
+  }
+
+  onBackdropKeyDown(event: KeyboardEvent): void {
+    if (event.key === 'Escape') {
+      this.closeModals();
+    }
+  }
+
+  saveApplication(): void {
+    this.currentApplication.statusText = PAYMENT_STATUS_LABELS[this.currentApplication.status];
+    if (this.isEditMode) {
+      const index = this.applications.findIndex((a) => a.id === this.currentApplication.id);
+      if (index !== -1) this.applications[index] = { ...this.currentApplication };
+    } else {
+      this.applications.push({ ...this.currentApplication });
+    }
+    this.closeModals();
+  }
+
+  deleteApplication(): void {
+    this.applications = this.applications.filter((a) => a.id !== this.currentApplication.id);
+    this.closeModals();
+  }
+
+  getTypeLabel(type: PaymentType): string {
+    return PAYMENT_TYPE_LABELS[type];
+  }
+
+  getMethodLabel(method?: PaymentMethod): string {
+    return method ? PAYMENT_METHOD_LABELS[method] : '—';
+  }
+
+  getStatusLabel(status: PaymentStatus): string {
+    return PAYMENT_STATUS_LABELS[status];
+  }
+
+  toggleColumnFilter(column: FilterColumn, event: MouseEvent): void {
     event.stopPropagation();
     if (this.activeFilterColumn === column) {
       this.activeFilterColumn = null;
+      this.filterMenuStyle = {};
       return;
     }
+
     this.activeFilterColumn = column;
     this.filterSearch = '';
     this.filterSortDraft = this.currentSortColumn === column ? this.currentSortDirection : '';
     const allValues = this.getUniqueColumnValues(column);
     const selected = this.columnSelectedValues[column];
     this.filterSelectedDraft = new Set(selected.length ? selected : allValues);
+
+    this.positionFilterMenu(event);
   }
 
-  closeColumnFilter(): void {
-    this.activeFilterColumn = null;
-  }
+  private positionFilterMenu(event: MouseEvent): void {
+    const trigger = event.target as HTMLElement;
+    const rect = trigger.getBoundingClientRect();
+    const menuWidth = 270;
+    const menuMaxHeight = 360;
+    const gap = 8;
 
-  setDraftSort(direction: 'asc' | 'desc'): void {
-    this.filterSortDraft = direction;
+    let left = rect.right - menuWidth;
+    let top = rect.bottom + gap;
+
+    if (left < gap) {
+      left = rect.left;
+    }
+    if (left + menuWidth > window.innerWidth - gap) {
+      left = window.innerWidth - menuWidth - gap;
+    }
+    if (top + menuMaxHeight > window.innerHeight - gap) {
+      top = rect.top - menuMaxHeight - gap;
+    }
+    if (top < gap) {
+      top = gap;
+    }
+
+    this.filterMenuStyle = {
+      left: `${Math.round(left)}px`,
+      top: `${Math.round(top)}px`,
+    };
   }
 
   clearCurrentColumnFilter(): void {
@@ -307,44 +419,62 @@ export class RequestAssessment implements OnInit, OnDestroy {
     this.currentPage = 1;
   }
 
+  setDraftSort(direction: 'asc' | 'desc'): void {
+    this.filterSortDraft = direction;
+  }
+
+  isSortDraftActive(direction: 'asc' | 'desc'): boolean {
+    return this.filterSortDraft === direction;
+  }
+
   get filterValues(): string[] {
     if (!this.activeFilterColumn) return [];
     const all = this.getUniqueColumnValues(this.activeFilterColumn);
     const term = this.filterSearch.trim().toLowerCase();
     if (!term) return all;
-    return all.filter((v) => v.toLowerCase().includes(term));
+    return all.filter((value) => value.toLowerCase().includes(term));
   }
 
   isValueChecked(value: string): boolean {
+    if (!this.activeFilterColumn) return false;
     return this.filterSelectedDraft.has(value);
+  }
+
+  toggleValue(value: string, checked: boolean): void {
+    if (checked) this.filterSelectedDraft.add(value);
+    else this.filterSelectedDraft.delete(value);
+  }
+
+  onToggleAllChange(event: Event): void {
+    const target = event.target as HTMLInputElement | null;
+    this.toggleAllValues(!!target?.checked);
+  }
+
+  onToggleValueChange(value: string, event: Event): void {
+    const target = event.target as HTMLInputElement | null;
+    this.toggleValue(value, !!target?.checked);
   }
 
   get isAllChecked(): boolean {
     const values = this.filterValues;
     if (!values.length) return false;
-    return values.every((v) => this.filterSelectedDraft.has(v));
+    return values.every((value) => this.isValueChecked(value));
   }
 
-  onToggleAllChange(event: Event): void {
-    const checked = !!(event.target as HTMLInputElement | null)?.checked;
-    for (const v of this.filterValues) {
-      if (checked) this.filterSelectedDraft.add(v);
-      else this.filterSelectedDraft.delete(v);
+  toggleAllValues(checked: boolean): void {
+    for (const value of this.filterValues) {
+      if (checked) this.filterSelectedDraft.add(value);
+      else this.filterSelectedDraft.delete(value);
     }
-  }
-
-  onToggleValueChange(value: string, event: Event): void {
-    const checked = !!(event.target as HTMLInputElement | null)?.checked;
-    if (checked) this.filterSelectedDraft.add(value);
-    else this.filterSelectedDraft.delete(value);
   }
 
   applyColumnFilter(): void {
     if (!this.activeFilterColumn) return;
     const column = this.activeFilterColumn;
-    const allValues = this.getUniqueColumnValues(column);
-    const selected = allValues.filter((v) => this.filterSelectedDraft.has(v));
-    this.columnSelectedValues[column] = selected.length === allValues.length ? [] : selected;
+    const values = this.getUniqueColumnValues(column);
+    const selected = values.filter((value) => this.filterSelectedDraft.has(value));
+    this.columnSelectedValues[column] = selected.length === values.length ? [] : selected;
+
     if (this.filterSortDraft) {
       this.currentSortColumn = column;
       this.currentSortDirection = this.filterSortDraft;
@@ -352,63 +482,72 @@ export class RequestAssessment implements OnInit, OnDestroy {
       this.currentSortColumn = null;
       this.currentSortDirection = '';
     }
-    this.closeColumnFilter();
-    this.applyFilters();
+
+    this.currentPage = 1;
+    this.activeFilterColumn = null;
   }
 
   cancelColumnFilter(): void {
-    this.closeColumnFilter();
+    this.activeFilterColumn = null;
   }
 
-  isSortDraftActive(direction: 'asc' | 'desc'): boolean {
-    return this.filterSortDraft === direction;
+  @HostListener('document:click')
+  onDocumentClick(): void {
+    this.activeFilterColumn = null;
   }
 
-  getCellValue(user: User, column: UserFilterColumn): string {
+  getCellValue(app: Application, column: FilterColumn): string {
     switch (column) {
-      case 'fullName':
-        return this.getFullName(user);
-      case 'email':
-        return user.email;
-      case 'phoneNumber':
-        return user.phoneNumber;
-      case 'role':
-        return this.getRoleLabel(user.role);
-      case 'isActive':
-        return user.isActive ? 'Активен' : 'Заблокирован';
-      case 'createdAt':
-        return this.formatDate(user.createdAt);
+      case 'id':
+        return String(app.id);
+      case 'date':
+        return app.date;
+      case 'sender':
+        return app.sender;
+      case 'recipient':
+        return app.recipient;
+      case 'type':
+        return this.getTypeLabel(app.type);
+      case 'amount':
+        return String(app.amount);
+      case 'fee':
+        return String(app.fee ?? 0);
+      case 'paymentMethod':
+        return this.getMethodLabel(app.paymentMethod);
+      case 'transactionId':
+        return app.transactionId || '—';
+      case 'receipt':
+        return app.receiptFileName || '—';
+      case 'statement':
+        return app.statementId || '—';
+      case 'status':
+        return this.getStatusLabel(app.status);
       case 'actions':
-        return '';
+        return 'Просмотр, Редактировать, Удалить';
       default:
         return '';
     }
   }
 
-  @HostListener('document:click')
-  onDocumentClick(): void {
-    this.closeColumnFilter();
-  }
-
-  private matchesColumnFilters(user: User): boolean {
-    for (const col of this.headerColumns.map((c) => c.key)) {
-      const selected = this.columnSelectedValues[col];
+  private matchesColumnFilters(app: Application): boolean {
+    for (const column of this.headerColumns.map((item) => item.key)) {
+      const selected = this.columnSelectedValues[column];
       if (!selected.length) continue;
-      const value = this.getCellValue(user, col);
+      const value = this.getCellValue(app, column);
       if (!selected.includes(value)) return false;
     }
     return true;
   }
 
-  private getUniqueColumnValues(column: UserFilterColumn): string[] {
+  private getUniqueColumnValues(column: FilterColumn): string[] {
     const values = new Set<string>();
-    for (const user of this.users) {
-      values.add(this.getCellValue(user, column));
+    for (const app of this.applications) {
+      values.add(this.getCellValue(app, column));
     }
     return Array.from(values).sort((a, b) => a.localeCompare(b, 'ru'));
   }
 
-  private compareByActiveSort(a: User, b: User): number {
+  private compareByActiveSort(a: Application, b: Application): number {
     if (!this.currentSortColumn || !this.currentSortDirection) return 0;
     const left = this.getCellValue(a, this.currentSortColumn);
     const right = this.getCellValue(b, this.currentSortColumn);
@@ -416,212 +555,22 @@ export class RequestAssessment implements OnInit, OnDestroy {
     return this.currentSortDirection === 'asc' ? result : -result;
   }
 
-  private updatePagination(): void {
-    this.totalPages = Math.ceil(this.filteredUsers.length / this.usersPerPage);
-    const start = (this.currentPage - 1) * this.usersPerPage;
-    this.paginatedUsers = this.filteredUsers.slice(start, start + this.usersPerPage);
-    this.buildPages();
-  }
-
-  private buildPages(): void {
-    this.pages = [];
-    for (let i = 1; i <= this.totalPages; i++) {
-      if (
-        i === 1 ||
-        i === this.totalPages ||
-        (i >= this.currentPage - 1 && i <= this.currentPage + 1)
-      ) {
-        this.pages.push(i);
-      } else if (i === this.currentPage - 2 || i === this.currentPage + 2) {
-        this.pages.push(-1);
-      }
-    }
-  }
-
-  changePage(page: number): void {
-    if (page < 1 || page > this.totalPages) return;
-    this.currentPage = page;
-    this.updatePagination();
-  }
-
-  // ========== ХЕЛПЕРЫ ==========
-
-  getFullName(user: User): string {
-    return `${user.lastName} ${user.firstName} ${user.middleName}`;
-  }
-
-  formatDate(dateString: string): string {
-    return new Date(dateString).toLocaleDateString('ru-RU', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  }
-
-  getRoleLabel(role: string): string {
-    return this.roleLabels[role] || role;
-  }
-
-  getRoleBadgeClass(role: string): string {
-    return this.roleBadgeClasses[role] || 'badge-primary';
-  }
-
-  // ========== НАВИГАЦИЯ ==========
-
-  viewUser(user: User): void {
-    this.selectedUser = user;
-    this.currentView = 'detail';
-  }
-
-  backToList(): void {
-    this.selectedUser = null;
-    this.currentView = 'list';
-  }
-
-  // ========== СОЗДАНИЕ / РЕДАКТИРОВАНИЕ ==========
-
-  openCreateModal(): void {
-    this.isEditing = false;
-    this.formData = { isActive: true };
-    this.formErrors = {};
-    this.showUserModal = true;
-  }
-
-  openEditModal(user: User): void {
-    this.isEditing = true;
-    this.formData = { ...user };
-    this.formErrors = {};
-    this.showUserModal = true;
-  }
-
-  editFromView(): void {
-    if (this.selectedUser) {
-      this.openEditModal(this.selectedUser);
-    }
-  }
-
-  closeUserModal(): void {
-    this.showUserModal = false;
-    this.formErrors = {};
-  }
-
-  saveUser(): void {
-    if (!this.validateForm()) return;
-
-    if (this.isEditing && this.formData.id) {
-      const index = this.users.findIndex((u) => u.id === this.formData.id);
-      if (index !== -1) {
-        this.users[index] = {
-          ...this.users[index],
-          ...(this.formData as User),
-          updatedAt: new Date().toISOString(),
-        };
-      }
-    } else {
-      const newUser: User = {
-        id: this.generateUUID(),
-        firstName: this.formData.firstName || '',
-        lastName: this.formData.lastName || '',
-        middleName: this.formData.middleName || '',
-        email: this.formData.email || '',
-        phoneNumber: this.formData.phoneNumber || '',
-        role: (this.formData.role as User['role']) || 'Applicant',
-        subscriptionPlan: (this.formData.subscriptionPlan as User['subscriptionPlan']) || 'Basic',
-        isActive: this.formData.isActive ?? true,
-        createdAt: new Date().toISOString(),
-        updatedAt: new Date().toISOString(),
-      };
-      this.users.push(newUser);
-    }
-
-    this.saveToStorage();
-    this.closeUserModal();
-    this.applyFilters();
-    this.currentView = 'list';
-  }
-
-  // ========== УДАЛЕНИЕ ==========
-
-  openDeleteModal(user: User): void {
-    this.userToDelete = user;
-    this.showDeleteModal = true;
-  }
-
-  deleteFromView(): void {
-    if (this.selectedUser) {
-      this.openDeleteModal(this.selectedUser);
-    }
-  }
-
-  closeDeleteModal(): void {
-    this.showDeleteModal = false;
-    this.userToDelete = null;
-  }
-
-  confirmDelete(): void {
-    if (!this.userToDelete) return;
-    const deleteId = this.userToDelete?.id;
-    this.users = this.users.filter((u) => u.id !== deleteId);
-    this.saveToStorage();
-    this.closeDeleteModal();
-    this.applyFilters();
-    this.currentView = 'list';
-    this.selectedUser = null;
-  }
-
-  // ========== ВАЛИДАЦИЯ ==========
-
-  private validateForm(): boolean {
-    this.formErrors = {};
-    let valid = true;
-
-    if (!this.formData.lastName?.trim()) {
-      this.formErrors['lastName'] = 'Поле обязательно для заполнения';
-      valid = false;
-    }
-    if (!this.formData.firstName?.trim()) {
-      this.formErrors['firstName'] = 'Поле обязательно для заполнения';
-      valid = false;
-    }
-    if (!this.formData.middleName?.trim()) {
-      this.formErrors['middleName'] = 'Поле обязательно для заполнения';
-      valid = false;
-    }
-
-    const email = this.formData.email?.trim() || '';
-    if (!email) {
-      this.formErrors['email'] = 'Поле обязательно для заполнения';
-      valid = false;
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)) {
-      this.formErrors['email'] = 'Введите корректный email';
-      valid = false;
-    } else {
-      const duplicate = this.users.find((u) => u.email === email && u.id !== this.formData.id);
-      if (duplicate) {
-        this.formErrors['email'] = 'Пользователь с таким email уже существует';
-        valid = false;
-      }
-    }
-
-    if (!this.formData.phoneNumber?.trim()) {
-      this.formErrors['phoneNumber'] = 'Поле обязательно для заполнения';
-      valid = false;
-    }
-    if (!this.formData.role) {
-      this.formErrors['role'] = 'Выберите роль';
-      valid = false;
-    }
-    if (!this.formData.subscriptionPlan) {
-      this.formErrors['subscriptionPlan'] = 'Выберите план подписки';
-      valid = false;
-    }
-
-    return valid;
-  }
-
-  onStatusToggle(): void {
-    this.formData.isActive = !this.formData.isActive;
+  goToPayments(): void {
+    this.router.navigate(['/admin', 'payments']);
   }
 }
+
+type FilterColumn =
+  | 'id'
+  | 'date'
+  | 'sender'
+  | 'recipient'
+  | 'type'
+  | 'amount'
+  | 'fee'
+  | 'paymentMethod'
+  | 'transactionId'
+  | 'receipt'
+  | 'statement'
+  | 'status'
+  | 'actions';
