--- libs/web/admin/src/lib/features/RequestAssessment/requests/requests.scss
@@ -1,363 +1,0 @@
-:host {
-  display: block;
-}
-
-.requests-container {
-  max-width: 1400px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.page-title {
-  margin-bottom: 2rem;
-  font-size: 1.8rem;
-  font-weight: 600;
-  color: var(--admin-text);
-}
-
-/* ========== ФИЛЬТРЫ ========== */
-.filters {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  margin-bottom: 2rem;
-}
-
-.search-box {
-  flex: 1;
-  min-width: 200px;
-  max-width: 400px;
-
-  input {
-    width: 100%;
-    padding: 0.75rem 1rem;
-    border: 1px solid var(--admin-border);
-    border-radius: var(--admin-radius-xs);
-
-    font-size: 0.95rem;
-
-    transition:
-      border-color 0.18s ease,
-      box-shadow 0.18s ease;
-
-    &:focus {
-      border-color: rgb(220 38 38 / 38%);
-      outline: none;
-      box-shadow: 0 0 0 4px rgb(220 38 38 / 10%);
-    }
-  }
-}
-
-.filter-select {
-  cursor: pointer;
-
-  min-width: 160px;
-  padding: 0.75rem 1rem;
-  border: 1px solid var(--admin-border);
-  border-radius: var(--admin-radius-xs);
-
-  font-size: 0.95rem;
-
-  background: var(--admin-surface);
-
-  transition:
-    border-color 0.18s ease,
-    box-shadow 0.18s ease;
-
-  &:focus {
-    border-color: rgb(220 38 38 / 38%);
-    outline: none;
-    box-shadow: 0 0 0 4px rgb(220 38 38 / 10%);
-  }
-}
-
-/* ========== ID CELL ========== */
-.id-cell {
-  font-family: monospace;
-  font-size: 0.85rem;
-  color: var(--admin-text-muted);
-}
-
-/* ========== BADGES ========== */
-.badge {
-  display: inline-block;
-
-  padding: 0.4rem 0.8rem;
-  border-radius: var(--admin-radius-round);
-
-  font-size: 0.8rem;
-  font-weight: 500;
-}
-
-.badge-success {
-  color: var(--admin-success-text);
-  background: var(--admin-success-bg);
-}
-
-.badge-danger {
-  color: var(--admin-critical-text);
-  background: var(--admin-critical-bg);
-}
-
-.badge-primary {
-  color: var(--admin-info-text);
-  background: var(--admin-info-bg);
-}
-
-.badge-warning {
-  color: var(--admin-warning-text);
-  background: var(--admin-warning-bg);
-}
-
-.badge-info {
-  color: #3730a3;
-  background: #e0e7ff;
-}
-
-/* ========== ПАГИНАЦИЯ ========== */
-.pagination {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  justify-content: center;
-
-  margin-top: 2rem;
-
-  button {
-    cursor: pointer;
-
-    padding: 0.5rem 1rem;
-    border: 1px solid var(--admin-border);
-    border-radius: 6px;
-
-    background: var(--admin-surface);
-
-    transition:
-      border-color 0.18s ease,
-      background-color 0.18s ease,
-      color 0.18s ease;
-
-    &:hover:not(:disabled) {
-      border-color: var(--admin-accent);
-      color: #fff;
-      background: var(--admin-accent);
-    }
-
-    &:disabled {
-      cursor: not-allowed;
-      opacity: 0.5;
-    }
-
-    &.active {
-      border-color: var(--admin-accent);
-      color: #fff;
-      background: var(--admin-accent);
-    }
-  }
-}
-
-/* ========== VIEW CARD ========== */
-.view-card {
-  max-width: 700px;
-  margin: 0 auto;
-}
-
-.view-card .view-details {
-  max-width: none;
-  margin: 0;
-}
-
-.view-card-actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75rem;
-  justify-content: flex-end;
-
-  margin-top: 1.25rem;
-  padding-top: 1.25rem;
-  border-top: 1px solid var(--admin-border);
-}
-
-/* ========== MODAL HEADER & CLOSE BTN ========== */
-.modal-header {
-  display: flex;
-  gap: 1rem;
-  align-items: flex-start;
-  justify-content: space-between;
-
-  margin-bottom: 1rem;
-
-  h3 {
-    margin: 0;
-  }
-}
-
-.close-btn {
-  cursor: pointer;
-
-  flex: 0 0 auto;
-
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  border: 1px solid var(--admin-border);
-  border-radius: 8px;
-
-  font-size: 1.25rem;
-  line-height: 1;
-  color: var(--admin-text-muted);
-
-  background: transparent;
-
-  &:hover {
-    border-color: var(--admin-accent);
-    color: var(--admin-accent);
-  }
-}
-
-/* ========== COLUMN FILTER ========== */
-.data-table {
-  th.filterable {
-    cursor: pointer;
-    user-select: none;
-    position: relative;
-  }
-
-  .no-wrap-col {
-    white-space: nowrap;
-  }
-
-  .header-cell {
-    display: flex;
-    gap: 0.4rem;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .filter-trigger {
-    cursor: pointer;
-
-    min-width: 22px;
-    height: 22px;
-    border: 1px solid var(--admin-accent);
-    border-radius: 4px;
-
-    font-size: 0.7rem;
-    line-height: 1;
-    color: #fff;
-
-    background: var(--admin-accent);
-
-    &:hover {
-      background: var(--admin-accent-hover, var(--admin-accent));
-    }
-  }
-
-  .column-filter-menu {
-    position: absolute;
-    z-index: 30;
-    top: 0;
-    left: calc(100% + 0.35rem);
-
-    overflow: auto;
-
-    padding: 0.65rem;
-    border: 1px solid var(--admin-border);
-    border-radius: 8px;
-
-    font-weight: 400;
-    color: var(--admin-text);
-
-    background: #fff;
-    box-shadow: 0 10px 22px rgb(0 0 0 / 16%);
-  }
-
-  .menu-section {
-    margin-bottom: 0.55rem;
-
-    input[type='text'] {
-      width: 100%;
-      padding: 0.35rem 0.5rem;
-      border: 1px solid var(--admin-border);
-      border-radius: 6px;
-
-      font-size: 0.85rem;
-      font-weight: 400;
-    }
-  }
-
-  .menu-btn {
-    cursor: pointer;
-
-    display: block;
-
-    width: 100%;
-    margin-bottom: 0.3rem;
-    padding: 0.4rem 0.45rem;
-    border: 1px solid var(--admin-border);
-    border-radius: 6px;
-
-    text-align: left;
-
-    background: #fff;
-
-    &:hover {
-      border-color: var(--admin-accent);
-      background: #eef3ff;
-    }
-
-    &.active {
-      border-color: var(--admin-accent);
-      color: #fff;
-      background: var(--admin-accent);
-    }
-  }
-
-  .clear-btn {
-    color: var(--admin-accent);
-  }
-
-  .value-list {
-    overflow: auto;
-
-    max-height: 145px;
-    padding: 0.35rem;
-    border: 1px solid var(--admin-border);
-    border-radius: 6px;
-  }
-
-  .checkbox-row {
-    display: flex;
-    gap: 0.45rem;
-    align-items: center;
-
-    margin-bottom: 0.25rem;
-
-    font-size: 0.85rem;
-    font-weight: 500;
-  }
-
-  .menu-actions {
-    display: flex;
-    gap: 0.45rem;
-    justify-content: flex-end;
-  }
-
-  .btn-sm {
-    padding: 0.3rem 0.7rem;
-    border-radius: 6px;
-    font-size: 0.8rem;
-  }
-}
-
-/* ========== АДАПТИВНОСТЬ ========== */
-@media (width <= 768px) {
-  .filters {
-    flex-direction: column;
-  }
-
-  .search-box {
-    width: 100%;
-    max-width: 100%;
-  }
-}
